settings: use Crowd for auth, for real this time.
