+++ liquid/settings.py
@@ -192,25 +192,10 @@ APPEND_SLASH = True
 
 SERVE_STATIC = True
 
-
-## LDAP STUFF ## should be in local_settings for security
-
-#AD_DNS_NAME='your-ldap-server.com'
-# If using non-SSL use these
-#AD_LDAP_PORT=389
-#AD_LDAP_URL='ldap://%s:%s' % (AD_DNS_NAME,AD_LDAP_PORT)
-# If using SSL use these:
-#AD_LDAP_PORT=636
-#AD_LDAP_URL='ldaps://%s:%s' % (AD_DNS_NAME,AD_LDAP_PORT)
-#AD_SEARCH_DN='dc=mygroup,dc=net,dc=com'
-#AD_NT4_DOMAIN='YOURDOMAIN'
-#AD_SEARCH_FIELDS= ['mail','givenName','sn','sAMAccountName','memberOf']
-#AD_MEMBERSHIP_REQ=['Group_Required','Alternative_Group']
-#AD_CERT_FILE='/path/to/your/cert.txt'
-#AD_DEBUG=True
-#AD_DEBUG_FILE='/path/to/writable/log/file/ldap.debug'
-AUTHENTICATION_BACKENDS = ('utils.ldapauth.ActiveDirectoryGroupMembershipSSLBackend',
-                           'django.contrib.auth.backends.ModelBackend',)
+AUTHENTICATION_BACKENDS = (
+  'utils.crowdauth.CrowdAuthenticationBackend',
+  'django.contrib.auth.backends.ModelBackend',
+)
 
 PASSWORD_HASHERS = (
     'django.contrib.auth.hashers.PBKDF2PasswordHasher',
@@ -243,16 +228,8 @@ LOGIN_REQUIRED_URLS = (
   (r'/corporate/resume/recruiter/(.*)$',False),
 )
 
-
-# ldap things
-AD_DNS_NAME= 'ad.uiuc.edu'
-AD_LDAP_PORT=636
-AD_LDAP_URL='ldaps://%s:%s' % (AD_DNS_NAME,AD_LDAP_PORT)
-AD_BIND_DN='cn={0},ou=Campus Accounts,dc=ad,dc=uiuc,dc=edu'
-AD_NT4_DOMAIN='AD.UILLINOIS.EDU'
-AD_SEARCH_FIELDS= ['givenName','sn','sAMAccountName']
-AD_DEBUG=True
-AD_DEBUG_FILE='/var/log/apache2/ldap.debug'
+CROWD_API_URL = 'http://engr-acm-sso-01.acm.illinois.edu:8095/crowd'
+CROWD_APP_NAME = 'acm-liquid'
 
 CRON_IPS = ['172.22.32.110']
 CRON_PASSWORD = ""
